requirements.txt
# Core
numpy==1.26.4
scipy==1.14.1
pandas==2.2.3

# PyTorch
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0

# NLP & Transformers
sentence-transformers==3.0.1
transformers==4.40.0

# ML
scikit-learn==1.5.2

# Vector Search
faiss-cpu==1.8.0
datasketch==1.6.5

# Web Framework
flask==3.0.3
flask-cors==5.0.0
werkzeug==3.0.4

# Document
python-docx==1.1.2

# Utilities
tqdm==4.66.5
psutil==6.0.0
matplotlib==3.9.2
requests==2.32.3
python-dotenv==1.0.1